+++ server/src/test/Review.test.ts
@@ -3,8 +3,8 @@ import axios from "axios";
 import { TokenPayload } from "google-auth-library";
 
 import { Review } from "common";
-import { Reviews, Students } from "../db/dbDefs";
-import * as Auth from "../endpoints/auth/routes";
+import { Reviews, Students } from "../../db/dbDefs";
+import * as Auth from "../api/auth/routes";
 import TestingServer from "./TestServer";
 
 const testingPort = 8080;
